+++ energyquantified/api/__init__.py
@@ -3,6 +3,7 @@ from .metadata import MetadataAPI
 from .timeseries import TimeseriesAPI
 from .periods import PeriodsAPI
 from .period_instances import PeriodInstancesAPI
+from .ohlc import OhlcAPI
 
 __all__ = [
     "InstancesAPI",
@@ -10,4 +11,5 @@ __all__ = [
     "TimeseriesAPI",
     "PeriodsAPI",
     "PeriodInstancesAPI",
+    "OhlcAPI",
 ]