+++ energyquantified/data/__init__.py
@@ -1,14 +1,21 @@
 from .timeseries import Timeseries, ValueType, Value, ScenariosValue, MeanScenariosValue
 from .periodseries import Periodseries, Period, CapacityPeriod
+from .ohlc import Product, OHLC, OHLCList
 
 
 __all__ = [
+    # Time series
     "Timeseries",
     "ValueType",
     "Value",
     "ScenariosValue",
     "MeanScenariosValue",
+    # Period series
     "Periodseries",
     "Period",
     "CapacityPeriod",
+    # OHLC
+    "Product",
+    "OHLC",
+    "OHLCList",
 ]